+++ ghost/admin/app/components/members/filters/email-open-rate.js
